+++ packages/server/src/server/rest-api.ts
@@ -4,8 +4,8 @@ import {
 } from "@anthem/utils";
 import axios from "axios";
 import express from "express";
+import { AxiosUtil, getHostFromNetworkName } from "../tools/axios-utils";
 import { logger } from "../tools/server-utils";
-import { AxiosUtil, getHostFromNetworkName } from "./axios-utils";
 import { PaginationParams } from "./resolvers/resolvers";
 import CELO from "./sources/celo";
 import COSMOS_EXTRACTOR from "./sources/cosmos-extractor";
