+++ packages/server/src/server/sources/celo.ts
@@ -5,7 +5,7 @@ import {
   IQuery,
   NetworkDefinition,
 } from "@anthem/utils";
-import { AxiosUtil, getHostFromNetworkName } from "../axios-utils";
+import { AxiosUtil, getHostFromNetworkName } from "../../tools/axios-utils";
 import { PaginationParams } from "../resolvers/resolvers";
 import { fetchProposalTitleFromID } from "./icarus-wings";
