+++ packages/server/src/server/sources/oasis.ts
@@ -6,8 +6,8 @@ import {
   IQuery,
   NetworkDefinition,
 } from "@anthem/utils";
+import { AxiosUtil, getHostFromNetworkName } from "../../tools/axios-utils";
 import { logSentryMessage } from "../../tools/server-utils";
-import { AxiosUtil, getHostFromNetworkName } from "../axios-utils";
 import { PaginationParams } from "../resolvers/resolvers";
 
 /** ===========================================================================
